Add +page.ts for loading data on pageload
